Move utilities to src folder
